requirements/local.txt
-r base.txt

django-debug-toolbar==1.9.1
factory-djoy==1.0.0
pytest-django==3.3.3
flake8==3.5.0